requirements.txt
Django==4.0.3
django-cors-headers
djangorestframework
requests